[REFACTOR] Update react structure about postlist
